+++ df-downloader-ui/src/components/tasks/task-type-icon.ts
@@ -1,6 +1,7 @@
 import DownloadIcon from "@mui/icons-material/Download";
 import SubtitlesIcon from "@mui/icons-material/Subtitles";
 import TaskIcon from "@mui/icons-material/Task";
+import MetadataIcon from "@mui/icons-material/DataObject";
 import { DfPipelineType } from "df-downloader-common";
 
 export const getTaskTypeIcon = (taskType: DfPipelineType) => {
@@ -9,6 +10,8 @@ export const getTaskTypeIcon = (taskType: DfPipelineType) => {
       return DownloadIcon;
     case "subtitles":
       return SubtitlesIcon;
+    case "update_download_meta":
+      return MetadataIcon;
     default:
       return TaskIcon;
   }